2020 update

+++ app/src/main/java/com/eduapps/edumage/oge_app/AudioTaskActivity.java
@@ -12,9 +12,9 @@ import android.media.AudioManager;
 import android.media.MediaPlayer;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.view.MenuItem;
@@ -555,7 +555,7 @@ public class AudioTaskActivity extends AppCompatActivity {
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         // the same behavior for the "up/home" button in the Action Bar
-        switch(item.getItemId()) {
+        switch (item.getItemId()) {
             case android.R.id.home:
                 releaseMediaPlayer();
                 this.onBackPressed();
@@ -765,7 +765,7 @@ public class AudioTaskActivity extends AppCompatActivity {
         db.insert(Tables.RecentActivities.TABLE_NAME, null, values);
 
         // also, update completion if user did the task well
-        if (rightAnswers / totalQuestions >= 0.6 && currentCompletion < 100) {
+        if ((double)rightAnswers / totalQuestions >= 0.6 && currentCompletion < 100) {
             if (currentCompletion + 50 == 100) {
                 SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(this);
                 SharedPreferences.Editor editor = preferences.edit();
